refactor(core): separate signature policy from codecs
